Initial dump of mcc's building block classes
I started writing a C-compiler called mcc which had a bunch of useful
classes for doing the menial things like reading files line by logical
line (according to the rules of C).  Those rules will likely be changed
(I'm not sure about line continuations, etc) but this is at least a base
library that compiles, links and passes the tests.