Chapter 16 - crafting interpreters
Scanning on demand.  Scans tokens for the lox language - I could modify
this, or I could modify my existing lexer for the token stream.  Dunno.
My way has a bunch of dynamic memory that would make it a *lot* slower,
and it uses full string comparison for the keywords, not a trie.