Native classes can be defined, instantiated and methods invoked
It's the culmination of a bunch of work, but native methods can be
invoked ob instances of Native classes!  I guess theoretically I could
add a regular class and define a native method on that, but I don't know
why I would... unless it were static?  Maybe?

Anyway, next stop is Native inheritance and then fleshing out the
standard library!