Get it to the point where I need to update the table
The table needs to exist and be usable prior to the stdlib being
instantiated (what do you think we use to register the classes, etc?).
So instead of using the stdlib hash, I'll update the table to take a
VALUE, which will allow me to use the stdlib String in the interpreter,
which allows for literal strings getting the same functionality as
dynamically instantiated ones.