Add >= and <= as explicit operations
It's cute and all, doing the <! trick, meaning fewer operations, but it
means we never explicitly call the >= operator meaning it couldn't be
overloaded.  I will probablly need to switch all this up to a comparable
in the future, so my users only have to do a single overload for
comparisons.