Remove the VM * scurge from the table
Now that we have accepted that the internal table can only have string
keys, we have removed the need to be able to allocate items during its
operation (which is a problem if we're iterating to garbage collect!).
This means that the VM * being passed around all over the place is
becoming much less necessary, but we'll chop the changes here for now.