Improve memory alloc/free for call_function
So, we can't delete _all_ objects associated with the VM, because some
of them might be shared, and one of them is a return value.  However, we
can get rid of a bunch more.  The main issue is the temp strings created
exactly for the purpose of printing, which still leaks memory.