Counter rethrown objects not previously thrown
I don't have to hurt performance of the exception handler stuff in a
major way if I only check at the end that we have a stacktrace which we
were already retrieving.  So rather than trust developers and have the
VM segfault and die, we will check and tell them that weirdness is
happening.