Fixes to make the release build work
GCC still has some fairly arcane error messages.  And yes, in a
technical sense it is correct - 0-1 would be a bad length of memory to
copy over when it is being cast to an unsigned number.  I just don't
believe that's a thing, you dipshit compiler that can't tell me the path
of code it went down to get there.