Add basic, brutal GC thread safety
If we attempt to allocate memory whilst collecting garbage, then it's a
pretty solid bet that we will free memory before it has had the chance
to be assigned to its object and Bad Things will happen.  This is only a
problem if you have multiple concurrent threads, which I intend to do,
soon.