Threads and Sockets work!
I am able to send and receive strings via a socket (TCP at least) and
all that within the same process, meaning that we have true OS-level
threads because I am making blocking calls in each thread that will stop
things, which would deadlock if that weren't a thing. Huzzah!