Add scaffolding for Module level imports
It doesn't do anything yet, but it's definitely the start of some
thoughts around module imports.  I should be able to limit true globals
to just the stdlib or "built in" functions.  I might even be able to
remove the compiler from the GC process, if I play my cards right.