Make the compiler a little more thread-safe
Remove another static variable, meaning that we no longer have shared
state between functions for one more piece.  So just the "main_thread"
(which will be the compiler thread) and we're good.  Oh, and the small
matter of marking the code with GC, I'll probably do stuff with the
module and a main function for that.