Improve the Windows 10 Port
Read the file with the Win32 API (because a HANDLE probably isn't
compatible with an FP).

Theoretically add in an untested Mutex implementation, but it's right
according to the docs, so I'm sure it will *just work*. Hur hur hur.

I still haven't implemented Condition Variables, because it's too much
for my brain right now.