Complete the Import..as implementation
Instead of having OP_IMPORT define the variable, we will leave that to a
later variable definition.  Further to that, we need to get rid of the
NIL on the top of the stack after the import function is called, which
means that a module is at the top of the stack and we can invoke
directly from that - which obvisouly requires an invoke implementation.