Use an enum for file open flags
I don't want users to have to learn C before they can effectively open a
file and write to it.  So instead, I've translated the API into an enum
instead of using a mode string.  This means I've also ditched some of
the flexibility of the mode string, but on purpose, because in my
experience, it's not used that much.  If you're opening a file for
writing, you probably want to create it.  And if you're opening a file
for reading, you probably don't.  And then there's that + on the
read_only param.  Don't get me started.