Use import for main file
I have the import code, which takes care of the full path from relative
thing and allows me to "circularly" import the "main" script without
re-compiling it or duplicating it in memory.  It's a weird use case for
sure, but programmers are weird people.  I should know.  Could get hairy
if I convert paths to "module names".