Fix multi-threaded GC
It really does need to be the whole of the realloc function that is the
critical section.  If we have one of the threads in GC and another
merrily allocating memory (which won't have been marked) then it's
probable that we'll free something we shouldn't, sooner or later. And we
know how much I enjoyed hunting the last bug :( Adding threads to that
can only make it more exciting.