Improved Windows compatibility
I haven't built on windows since including C++ in the project, so there
were some tweaks to be made there, and I've recently discovered that the
MSVC compiler does support standards other than C89, but they're not the
default, so need to be requested.

This also means I've improved the Windows file support now that I have
the flags.