Allow stack printing from userland
Toggle stack printing on and off from comet code if the build has debug
trace enabled.  Most of the time I don't care about the preamble and I
just want to start at a specific spot.  So I'm going to allow myself to
switch it on and off from userland.  Makes sense, right?