Improve the odds of call_function results living
In the hopes of not having the result of a function call being GC'd
between the call and return doing something with it, I have hacked in
'marking' it safe from collection for the next round.  It means we will
have a (potentially) larger working memory footprint than strictly
necessary, but that's better than having a slightly smaller footprint
than strictly necessary.