Shift mark stdlib obj into definition
If the object in the stdlib contains other objects, then it needs a
special function to 'mark' it as in use, such that it marks the objects
it references.  Instead of making this special to the stdlib, just make
it part of the native class definition, such that the vm needs less
knowledge about the type of stdlib object in order to mark it.