Fix to have clean run of is.cmt unit test
Some string builder fixes, some parser fixes for import and some
re-arrangements of the unit test output.  Either way, the easy test
runs.  string_test.cmt does _not_ run but I suspect some other construct
is leaving the stack in a bad way and that's causing unhappiness when
the file is run as a dynamic import.