Convert two more scripts to unit tests
Now that I have a unit test framework, I want to be able to run these
tests for each change and figure out if something has gone wrong.

The Socket test needed to be disabled and the exception thread needed to
not throw, and especially not throw on import.  I should probably add
something like 'Assert.that(x).throws(Exception)' for better testing.