Convert coverage output to comet
The output was getting very unwieldy to do in c++, but could be fairly
effectively done in comet itself.  So I've shifted a bunch of it to
comet.  Of course, to do that, I had to shave a few yaks.  It's not
finished, on account of the module::get_execution_counts isn't
completed.